@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "format": "prettier --write \"src/**/*.{ts,tsx,css}\""
   },
   "dependencies": {
     "@clerk/nextjs": "^4.25.1",
@@ -16,25 +17,28 @@
     "@prisma/client": "^5.2.0",
     "@sanity/image-url": "^1.0.2",
     "@sanity/vision": "^3.15.1",
+    "@types/jsonwebtoken": "^9.0.3",
     "@types/react": "18.2.20",
-    "@types/react-dom": "18.2.7",
+    "@types/react-dom": "^18.2.12",
     "autoprefixer": "10.4.15",
     "axios": "^1.5.0",
     "bufferutil": "^4.0.7",
     "date-fns": "^2.30.0",
-    "eslint-config-next": "13.4.19",
+    "eslint-config-next": "^13.5.4",
     "framer-motion": "^10.16.0",
     "gsap": "^3.12.2",
     "install": "^0.13.0",
     "jsonwebtoken": "^9.0.2",
     "locomotive-scroll": "^4.1.4",
     "lottie-react": "^2.4.0",
-    "next": "13.4.19",
+    "next": "^13.5.4",
+    "next-nprogress-bar": "^2.1.2",
     "next-sanity": "^5.4.4",
     "next-themes": "^0.2.1",
-    "postcss": "8.4.28",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "postcss": "^8.4.31",
+    "prettier": "^3.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-toastify": "^9.1.3",
     "sanity": "^3.15.1",
@@ -42,14 +46,15 @@
     "styled-components": "5.2",
     "tailwindcss": "3.3.3",
     "utf-8-validate": "^5.0.10",
-    "uuid": "^9.0.0",
-    "zod": "^3.22.2"
+    "uuid": "^9.0.1",
+    "zod": "^3.22.2",
+    "zustand": "^4.4.3"
   },
   "devDependencies": {
-    "@types/node": "^20.5.9",
+    "@types/node": "^20.8.4",
     "@typescript-eslint/eslint-plugin": "^6.6.0",
     "@typescript-eslint/parser": "^6.6.0",
-    "eslint": "^8.48.0",
+    "eslint": "^8.51.0",
     "eslint-plugin-jsdoc": "^46.5.1",
     "prisma": "^5.2.0",
     "ts-node": "^10.9.1",
