@@ -1,21 +1,31 @@
-const { v4: uuidv4 } = require('uuid');
+const {v4: uuidv4} = require('uuid');
 
 function isValidEmail(email) {
-    // Basic email validation regex
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    return emailRegex.test(email);
-  }
+  // Basic email validation regex
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return emailRegex.test(email);
+}
 
 function validateLoginInput(name, email, password) {
   const errors = [];
 
   // Validate email
-  if (!email || typeof email !== 'string' || email.length > 35 || !isValidEmail(email)) {
+  if (
+    !email ||
+    typeof email !== 'string' ||
+    email.length > 35 ||
+    !isValidEmail(email)
+  ) {
     errors.push('Email must be a valid string up to 35 characters.');
   }
 
   // Validate password
-  if (!password || typeof password !== 'string' || password.length < 6 || password.length > 32) {
+  if (
+    !password ||
+    typeof password !== 'string' ||
+    password.length < 6 ||
+    password.length > 32
+  ) {
     errors.push('Password must be a string with 6 to 32 characters.');
   }
 
@@ -23,9 +33,9 @@ function validateLoginInput(name, email, password) {
 }
 
 // Example usage
-const name = "John Doe";
-const email = "john.doe@example.com";
-const password = "mysecretpassword";
+const name = 'John Doe';
+const email = 'john.doe@example.com';
+const password = 'mysecretpassword';
 
 const validationErrors = validateLoginInput(name, email, password);
 
@@ -37,5 +47,5 @@ if (validationErrors.length === 0) {
 }
 
 module.exports = {
-    validateLoginInput,
-}
+  validateLoginInput,
+};
