@@ -1,20 +1,37 @@
-const { v4: uuidv4 } = require('uuid');
+const {v4: uuidv4} = require('uuid');
 
 function validateSignupInput(name, email, password) {
   const errors = [];
 
   // Validate name
-  if (!name || typeof name !== 'string' || name.length > 35 || /\d/.test(name)) {
-    errors.push('Name must be a string without numerics and up to 35 characters.');
+  if (
+    !name ||
+    typeof name !== 'string' ||
+    name.length > 35 ||
+    /\d/.test(name)
+  ) {
+    errors.push(
+      'Name must be a string without numerics and up to 35 characters.'
+    );
   }
 
   // Validate email
-  if (!email || typeof email !== 'string' || email.length > 35 || !isValidEmail(email)) {
+  if (
+    !email ||
+    typeof email !== 'string' ||
+    email.length > 35 ||
+    !isValidEmail(email)
+  ) {
     errors.push('Email must be a valid string up to 35 characters.');
   }
 
   // Validate password
-  if (!password || typeof password !== 'string' || password.length < 6 || password.length > 32) {
+  if (
+    !password ||
+    typeof password !== 'string' ||
+    password.length < 6 ||
+    password.length > 32
+  ) {
     errors.push('Password must be a string with 6 to 32 characters.');
   }
 
@@ -28,9 +45,9 @@ function isValidEmail(email) {
 }
 
 // Example usage
-const name = "John Doe";
-const email = "john.doe@example.com";
-const password = "mysecretpassword";
+const name = 'John Doe';
+const email = 'john.doe@example.com';
+const password = 'mysecretpassword';
 
 const validationErrors = validateSignupInput(name, email, password);
 
@@ -42,5 +59,5 @@ if (validationErrors.length === 0) {
 }
 
 module.exports = {
-    validateSignupInput,
-}
+  validateSignupInput,
+};
