@@ -7,9 +7,9 @@ import {socket} from './socket';
  * This is a chat component
  * @return {JSX.Element}
  */
-const Chat = ():JSX.Element => {
+const Chat = (): JSX.Element => {
   const [message, setMessage] = useState<string>('');
-  const [ messages, setMessages] = useState<Array | []>([]);
+  const [messages, setMessages] = useState<Array | []>([]);
   const username = 'test';
 
   /**
@@ -18,7 +18,7 @@ const Chat = ():JSX.Element => {
    * @param {string} username
    * @return {Promise<void>}
    */
-  async function sendMessage(username:string, message:string): Promise<void> {
+  async function sendMessage(username: string, message: string): Promise<void> {
     await socket.emit('chat message', {user: username, message});
   }
   // get the messages from the server
@@ -37,37 +37,37 @@ const Chat = ():JSX.Element => {
       console.log(message);
       if (message !== null) sendMessage(username, message);
       // setMessages([...messages, { user: username, message: message }]);
-      setMessage("");
+      setMessage('');
     };
   }, [message, username]);
 
   return (
-    <div className="mx-10 h-[100vh]">
-      <div className="flex flex-col y-overflow-auto bg-slate-400 ">
+    <div className='mx-10 h-[100vh]'>
+      <div className='flex flex-col y-overflow-auto bg-slate-400 '>
         {messages.map((message, index) => (
-          <div key={index} className="flex flex-col">
+          <div key={index} className='flex flex-col'>
             <p>{message.user}: </p>
-            <p className="ml-5">{message.message}</p>
+            <p className='ml-5'>{message.message}</p>
           </div>
         ))}
       </div>
       <br />
       <br />
 
-      <div className="flex flex-row fixed bottom-0 w-[93.5vw] mx-1">
+      <div className='flex flex-row fixed bottom-0 w-[93.5vw] mx-1'>
         <input
-          id="input"
+          id='input'
           value={message}
           onChange={(e) => setMessage(e.target.value)}
-          type="text"
-          className="w-full h-8 p-2 rounded-md border border-gray-300 focus:outline-none focus:border-blue-500"
-          placeholder="Enter text..."
+          type='text'
+          className='w-full h-8 p-2 rounded-md border border-gray-300 focus:outline-none focus:border-blue-500'
+          placeholder='Enter text...'
         />
         <button
-          id="button"
-          defaultValue={" "}
+          id='button'
+          defaultValue={' '}
           onClick={setShow}
-          className="h-8 w-20 p-1 rounded-md bg-blue-500 text-white focus:outline-none focus:bg-blue-600"
+          className='h-8 w-20 p-1 rounded-md bg-blue-500 text-white focus:outline-none focus:bg-blue-600'
         >
           Submit
         </button>
