@@ -1,7 +1,6 @@
-import { io } from "socket.io-client";
+import {io} from 'socket.io-client';
 
-const URL:string =
-  process.env.NODE_ENV === "production" ? undefined : "http://localhost:3200";
+const URL: string = process.env.NODE_ENV === 'production' ? undefined : 'http://localhost:3200';
 
 export const socket = io(URL, {
   autoConnect: false,
