@@ -1,11 +1,11 @@
 'use server';
-import { cookies } from 'next/headers'
-interface Token{
-    name?: string;
-    value?: string;
+import {cookies} from 'next/headers';
+interface Token {
+  name?: string;
+  value?: string;
 }
 export default async function handler() {
-    const cookieStore = cookies();
-    const token: Token = cookieStore.get('token');
-    return token.value;
+  const cookieStore = cookies();
+  const token: Token = cookieStore.get('token') || {name: '', value: ''};
+  return token.value || '';
 }
