@@ -1,6 +1,6 @@
 'use client';
-import React from "react";
-import { Spinner } from "@nextui-org/react";
+import React from 'react';
+import {Spinner} from '@nextui-org/react';
 import Posts from '@/components/Cards/Posts';
 //import { ThemeSwitcher } from '@/components/Buttons/ThemeSwitcher';
 
@@ -20,14 +20,15 @@ type PostProps = {
 export default function Home() {
   const [posts, setPosts] = React.useState([]);
   React.useEffect(() => {
-    fetch(`http://localhost:3500/status/command/posts/get`)
-      .then(res => res.json())
-      .then(data => {
+    fetch(`http://localhost:3500/status/command/posts/get`, {revalidate: 20})
+      .then((res) => res.json())
+      .then((data) => {
         setPosts(data);
         console.log(data);
       })
-      .catch(err => console.log(err));
+      .catch((err) => console.log(err));
   }, []);
+  console.log(posts);
 
   return (
     <center>
@@ -38,12 +39,15 @@ export default function Home() {
           <div key={post.id}>
             {index % 3 === 0 ? (
               <div key={post.id}>
-              <div aria-hidden='true' className='bg-layers bg-scale w-56 h-56 m-auto blur-xl bg-gradient-to-r dark:from-purple-800 dark:via-blue-700 dark:to-cyan-600 from-violet-300 via-cyan-300 to-purple-300 rounded-full md:w-[45rem] md:h-[25rem] md:blur-3xl absolute'></div>
-              <Posts {...post}/>
+                <div
+                  aria-hidden='true'
+                  className='bg-layers bg-scale w-56 h-56 m-auto blur-xl bg-gradient-to-r dark:from-purple-800 dark:via-blue-700 dark:to-cyan-600 from-violet-300 via-cyan-300 to-purple-300 rounded-full md:w-[45rem] md:h-[25rem] md:blur-3xl absolute'
+                ></div>
+                <Posts {...post} />
               </div>
             ) : (
               <div key={post.id}>
-                <Posts {...post}/>
+                <Posts {...post} />
               </div>
             )}
           </div>
