@@ -1,4 +1,4 @@
-import { authMiddleware } from "@clerk/nextjs";
+import {authMiddleware} from '@clerk/nextjs';
 
 // This example protects all routes including api/trpc routes
 // Please edit this to allow other routes to be public as needed.
@@ -6,5 +6,5 @@ import { authMiddleware } from "@clerk/nextjs";
 export default authMiddleware({});
 
 export const config = {
-      matcher: ['/((?!.+\\.[\\w]+$|_next).*)', '/', '/(api|trpc)(.*)'],
+  matcher: ['/((?!.+\\.[\\w]+$|_next).*)', '/', '/(api|trpc)(.*)'],
 };
