@@ -1,30 +1,23 @@
 'use client';
-import React from "react";
+import React from 'react';
 
-import {Card, CardHeader, CardBody, CardFooter, Divider, Link, Image} from "@nextui-org/react";
+import {Card, CardHeader, CardBody, CardFooter, Divider} from '@nextui-org/react';
 
 export default function Intro(data) {
-    console.log(data)
+  console.log(data);
   return (
-    <Card className="max-w-[400px]">
-      <CardHeader className="flex gap-3">
-        <p className="text-md text-bold">NextUI</p>
+    <Card className='max-w-[400px]'>
+      <CardHeader className='flex gap-3'>
+        <p className='text-md text-bold'>NextUI</p>
       </CardHeader>
-      <Divider/>
+      <Divider />
       <CardBody>
-        <div className='flex flex-col'>
-            workplaces first
-        </div>
-        <div className='flex flex-col'>
-            education second
-        </div>
-        <div className='flex flex-col'>
-            contacts third
-        </div>
+        <div className='flex flex-col'>workplaces first</div>
+        <div className='flex flex-col'>education second</div>
+        <div className='flex flex-col'>contacts third</div>
       </CardBody>
-      <Divider/>
-      <CardFooter>
-      </CardFooter>
+      <Divider />
+      <CardFooter></CardFooter>
     </Card>
   );
 }
