@@ -1,4 +1,4 @@
-'use client'
+'use client';
 
 /**
  * This route is responsible for the built-in authoring environment using Sanity Studio.
@@ -9,9 +9,9 @@
  * https://github.com/sanity-io/next-sanity
  */
 
-import { NextStudio } from 'next-sanity/studio'
-import config from '../../../../sanity.config'
+import {NextStudio} from 'next-sanity/studio';
+import config from '../../../../sanity.config';
 
 export default function StudioPage() {
-  return <NextStudio config={config} />
+  return <NextStudio config={config} />;
 }
