@@ -1,18 +1,18 @@
 // app/components/ThemeSwitcher.tsx
-"use client";
+'use client';
 
-import {useTheme} from "next-themes";
-import { useEffect, useState } from "react";
+import {useTheme} from 'next-themes';
+import {useEffect, useState} from 'react';
 
 export function ThemeSwitcher() {
-  const [mounted, setMounted] = useState(false)
-  const { theme, setTheme } = useTheme()
+  const [mounted, setMounted] = useState(false);
+  const {theme, setTheme} = useTheme();
 
   useEffect(() => {
-    setMounted(true)
-  }, [])
+    setMounted(true);
+  }, []);
 
-  if(!mounted) return null
+  if (!mounted) return null;
 
   return (
     <div>
@@ -20,5 +20,5 @@ export function ThemeSwitcher() {
       <button onClick={() => setTheme('light')}>Light Mode</button>
       <button onClick={() => setTheme('dark')}>Dark Mode</button>
     </div>
-  )
-};
+  );
+}
